+++ src/Frozennode/Administrator/DataTable/Columns/Column.php
@@ -10,21 +10,21 @@ class Column {
 	/**
 	 * The validator instance
 	 *
-	 * @var Frozennode\Administrator\Validator
+	 * @var \Frozennode\Administrator\Validator
 	 */
 	protected $validator;
 
 	/**
 	 * The config instance
 	 *
-	 * @var Frozennode\Administrator\Config\ConfigInterface
+	 * @var \Frozennode\Administrator\Config\ConfigInterface
 	 */
 	protected $config;
 
 	/**
 	 * The config instance
 	 *
-	 * @var Illuminate\Database\DatabaseManager
+	 * @var \Illuminate\Database\DatabaseManager
 	 */
 	protected $db;
 
@@ -59,6 +59,7 @@ class Column {
 		'is_included' => false,
 		'external' => false,
 		'belongs_to_many' => false,
+		'visible' => true,
 	);
 
 	/**
@@ -105,9 +106,9 @@ class Column {
 	/**
 	 * Create a new action Factory instance
 	 *
-	 * @param Frozennode\Administrator\Validator 				$validator
-	 * @param Frozennode\Administrator\Config\ConfigInterface	$config
-	 * @param Illuminate\Database\DatabaseManager 				$db
+	 * @param \Frozennode\Administrator\Validator 				$validator
+	 * @param \Frozennode\Administrator\Config\ConfigInterface	$config
+	 * @param \Illuminate\Database\DatabaseManager 				$db
 	 * @param array												$options
 	 */
 	public function __construct(Validator $validator, ConfigInterface $config, DB $db, array $options)
@@ -171,6 +172,14 @@ class Column {
 		else
 		{
 			$options['is_included'] = true;
+		}
+
+		//run the visible property closure if supplied
+		$visible = $this->validator->arrayGet($options, 'visible');
+
+		if (is_callable($visible))
+		{
+			$options['visible'] = $visible($this->config->getDataModel()) ? true : false;
 		}
 
 		$this->suppliedOptions = $options;