+++ changelog.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### 4.6.0
+- Support for smaller screens and mobile devices
+- Visible option for columns that accepts either a boolean or closure
+- Relationship constraints now work with hasMany and hasOne fields
+- There is now an `options_filter` option for relationship fields that lets you modify the query before getting the relationship options
+- Custom actions and saves now rebuild the supplied config file after performing the action
+- The `editable` property now accepts a closure and is passed the current page's data object
+- New translations (da, it)
+- Bugfix: Constraint fields no longer make multiple requests at a single time
+- Bugfix: The key field is no longer set on models. This would cause some bugs on some setups
+
 ### 4.5.0
 - You can now provide `value`, `min_value`, and `max_value` options in filter fields to set default values
 - It is now possible to specify as many submenus in the `menus` array as you want